Keep background priming lifecycle state in one owner
Background workspace priming was split between ContentView, a waiter, and task-local retry counters. Moving the lifecycle into a dedicated MainActor coordinator gives the retry ledger the same lifetime as the view-owned task and removes the subscription retain cycle that made waiter deinit cleanup unreliable.

Constraint: Do not run xcodebuild locally; validation must avoid the prohibited build path.

Rejected: Keep timeout counts local to primePendingBackgroundWorkspaces | SwiftUI task-id restarts reset local state when pending IDs change.

Confidence: medium

Scope-risk: moderate

Directive: Keep background priming ownership in BackgroundWorkspacePrimeCoordinator; ContentView should only start the task.

Tested: git diff --cached --check; swiftc -parse Sources/BackgroundWorkspacePrimeCoordinator.swift

Not-tested: Full app build and CI before push.
