+++ Sources/ContentView.swift
@@ -1602,6 +1602,7 @@ struct ContentView: View {
     @StateObject private var fileExplorerStore = FileExplorerStore()
     @StateObject private var sessionIndexStore = SessionIndexStore()
     @StateObject private var selectedWorkspaceDirectoryObserver = SelectedWorkspaceDirectoryObserver()
+    @State private var backgroundWorkspacePrimeCoordinator = BackgroundWorkspacePrimeCoordinator()
     @State private var fileExplorerWidth: CGFloat = 220
     @State private var fileExplorerDragStartWidth: CGFloat?
     @State private var previousSelectedWorkspaceId: UUID?
@@ -3281,8 +3282,8 @@ struct ContentView: View {
 
         // Prime background workspaces off-screen. Rendering them just to run a task
         // mounts every keepAllAlive tab view and can materialize hidden terminals.
-        view = AnyView(view.task(id: pendingBackgroundWorkspaceLoadTaskKey) {
-            await primePendingBackgroundWorkspaces()
+        view = AnyView(view.task(id: backgroundWorkspacePrimeCoordinator.taskKey(for: tabManager)) {
+            await backgroundWorkspacePrimeCoordinator.primePendingBackgroundWorkspaces(tabManager: tabManager)
         })
 
         view = AnyView(view.onReceive(tabManager.$debugPinnedWorkspaceLoadIds) { _ in
@@ -3847,223 +3848,6 @@ struct ContentView: View {
             }
         }
 #endif
-    }
-
-    private enum BackgroundWorkspacePrimeState {
-        case pending
-        case completed(reason: String)
-    }
-
-    private enum BackgroundWorkspacePrimePolicy {
-        static let timeoutSeconds: TimeInterval = 2.0
-        static let maxTimeoutRetries = 3
-    }
-
-    private final class BackgroundWorkspacePrimeWaiter: @unchecked Sendable {
-        private let lock = NSLock()
-        private var continuation: CheckedContinuation<String, Never>?
-        private var cleanupActions: [() -> Void] = []
-        private var resolvedReason: String?
-
-        var isResolved: Bool { lock.lock(); defer { lock.unlock() }; return resolvedReason != nil }
-
-        deinit { finish(reason: "cancelled") }
-
-        func start(continuation: CheckedContinuation<String, Never>) {
-            let reason: String?
-            lock.lock(); reason = resolvedReason; if reason == nil { self.continuation = continuation }; lock.unlock()
-            if let reason { continuation.resume(returning: reason) }
-        }
-
-        func addCancellable(_ cancellable: AnyCancellable) { addCleanup { cancellable.cancel() } }
-        func addObserver(_ observer: NSObjectProtocol) { addCleanup { NotificationCenter.default.removeObserver(observer) } }
-        func addTimeoutWorkItem(_ workItem: DispatchWorkItem) { addCleanup { workItem.cancel() } }
-
-        func finish(reason: String) {
-            let drained: (CheckedContinuation<String, Never>?, [() -> Void])?
-            lock.lock()
-            if resolvedReason == nil { resolvedReason = reason; drained = (continuation, cleanupActions); continuation = nil; cleanupActions.removeAll() } else { drained = nil }
-            lock.unlock()
-            guard let (continuation, cleanupActions) = drained else { return }
-            cleanupActions.forEach { $0() }; continuation?.resume(returning: reason)
-        }
-
-        private func addCleanup(_ action: @escaping () -> Void) {
-            lock.lock()
-            guard resolvedReason == nil else { lock.unlock(); action(); return }
-            cleanupActions.append(action); lock.unlock()
-        }
-    }
-
-    private func primeBackgroundWorkspaceIfNeeded(workspaceId: UUID) async -> String {
-        let shouldPrime = await MainActor.run {
-            tabManager.pendingBackgroundWorkspaceLoadIds.contains(workspaceId)
-        }
-        guard shouldPrime else { return "already_cleared" }
-
-#if DEBUG
-        let startedAt = ProcessInfo.processInfo.systemUptime
-        cmuxDebugLog("workspace.backgroundPrime.start workspace=\(workspaceId.uuidString.prefix(5))")
-#endif
-
-        let initialState = await MainActor.run {
-            stepBackgroundWorkspacePrime(workspaceId: workspaceId)
-        }
-        let completionReason: String
-        switch initialState {
-        case .completed(let reason):
-            completionReason = reason
-        case .pending:
-            completionReason = await waitForBackgroundWorkspacePrimeCompletion(
-                workspaceId: workspaceId,
-                timeoutSeconds: BackgroundWorkspacePrimePolicy.timeoutSeconds
-            )
-        }
-#if DEBUG
-        let elapsedMs = (ProcessInfo.processInfo.systemUptime - startedAt) * 1000
-        cmuxDebugLog(
-            "workspace.backgroundPrime.finish workspace=\(workspaceId.uuidString.prefix(5)) " +
-            "reason=\(completionReason) ms=\(String(format: "%.2f", elapsedMs))"
-        )
-#endif
-        return completionReason
-    }
-
-    @MainActor
-    private func stepBackgroundWorkspacePrime(workspaceId: UUID) -> BackgroundWorkspacePrimeState {
-        guard tabManager.pendingBackgroundWorkspaceLoadIds.contains(workspaceId) else {
-            return .completed(reason: "already_cleared")
-        }
-        guard let workspace = tabManager.tabs.first(where: { $0.id == workspaceId }) else {
-            tabManager.completeBackgroundWorkspaceLoad(for: workspaceId)
-            return .completed(reason: "workspace_removed")
-        }
-
-        workspace.requestBackgroundPrimeTerminalSurfaceStartIfNeeded()
-        guard workspace.hasLoadedBackgroundPrimeTerminalSurface() else {
-            return .pending
-        }
-
-        tabManager.completeBackgroundWorkspaceLoad(for: workspaceId)
-        return .completed(reason: "surface_ready")
-    }
-
-    @MainActor
-    private func waitForBackgroundWorkspacePrimeCompletion(
-        workspaceId: UUID,
-        timeoutSeconds: TimeInterval
-    ) async -> String {
-        let waiter = BackgroundWorkspacePrimeWaiter()
-        return await withTaskCancellationHandler {
-            await withCheckedContinuation { (continuation: CheckedContinuation<String, Never>) in
-                Task { @MainActor in
-                    waiter.start(continuation: continuation)
-                    guard !waiter.isResolved else { return }
-
-                    @MainActor
-                    func evaluate() {
-                        switch stepBackgroundWorkspacePrime(workspaceId: workspaceId) {
-                        case .pending:
-                            break
-                        case .completed(let reason):
-                            waiter.finish(reason: reason)
-                        }
-                    }
-
-                    waiter.addCancellable(
-                        tabManager.$pendingBackgroundWorkspaceLoadIds
-                            .map { _ in () }
-                            .sink { _ in
-                                Task { @MainActor in evaluate() }
-                            }
-                    )
-                    waiter.addCancellable(
-                        tabManager.$tabs
-                            .map { _ in () }
-                            .sink { _ in
-                                Task { @MainActor in evaluate() }
-                            }
-                    )
-                    if let workspace = tabManager.tabs.first(where: { $0.id == workspaceId }) {
-                        waiter.addCancellable(
-                            workspace.$panels
-                                .map { _ in () }
-                                .sink { _ in
-                                    Task { @MainActor in evaluate() }
-                                }
-                        )
-                    }
-
-                    let readyObserver = NotificationCenter.default.addObserver(
-                        forName: .terminalSurfaceDidBecomeReady,
-                        object: nil,
-                        queue: .main
-                    ) { notification in
-                        guard let readyWorkspaceId = notification.userInfo?["workspaceId"] as? UUID,
-                              readyWorkspaceId == workspaceId else { return }
-                        Task { @MainActor in evaluate() }
-                    }
-                    waiter.addObserver(readyObserver)
-
-                    let hostedViewObserver = NotificationCenter.default.addObserver(
-                        forName: .terminalSurfaceHostedViewDidMoveToWindow,
-                        object: nil,
-                        queue: .main
-                    ) { notification in
-                        guard let readyWorkspaceId = notification.userInfo?["workspaceId"] as? UUID,
-                              readyWorkspaceId == workspaceId else { return }
-                        Task { @MainActor in evaluate() }
-                    }
-                    waiter.addObserver(hostedViewObserver)
-
-                    let timeoutWorkItem = DispatchWorkItem {
-                        Task { @MainActor in
-                            if case .completed(let reason) = stepBackgroundWorkspacePrime(workspaceId: workspaceId) {
-                                waiter.finish(reason: reason)
-                            } else {
-                                waiter.finish(reason: "timeout")
-                            }
-                        }
-                    }
-                    waiter.addTimeoutWorkItem(timeoutWorkItem)
-                    DispatchQueue.main.asyncAfter(
-                        deadline: .now() + timeoutSeconds,
-                        execute: timeoutWorkItem
-                    )
-
-                    evaluate()
-                }
-            }
-        } onCancel: {
-            Task { @MainActor in
-                waiter.finish(reason: "cancelled")
-            }
-        }
-    }
-
-    private var pendingBackgroundWorkspaceLoadTaskKey: [String] {
-        tabManager.pendingBackgroundWorkspaceLoadIds
-            .map(\.uuidString)
-            .sorted()
-    }
-
-    private func primePendingBackgroundWorkspaces() async {
-        var timeoutCounts: [UUID: Int] = [:]
-        while !Task.isCancelled {
-            let workspaceIds = await MainActor.run { tabManager.pendingBackgroundWorkspaceLoadIds.sorted { $0.uuidString < $1.uuidString } }
-            guard !workspaceIds.isEmpty else { return }
-            for workspaceId in workspaceIds {
-                guard !Task.isCancelled else { return }
-                let reason = await primeBackgroundWorkspaceIfNeeded(workspaceId: workspaceId)
-                guard reason == "timeout" else { timeoutCounts[workspaceId] = nil; continue }
-                let timeoutCount = (timeoutCounts[workspaceId] ?? 0) + 1
-                timeoutCounts[workspaceId] = timeoutCount
-                if timeoutCount >= BackgroundWorkspacePrimePolicy.maxTimeoutRetries {
-                    await MainActor.run { tabManager.completeBackgroundWorkspaceLoad(for: workspaceId) }
-                    timeoutCounts[workspaceId] = nil
-                }
-            }
-        }
     }
 
     private func addTab() {
